requirements1.txt
alembic==0.9.9
altgraph==0.15
aniso8601==3.0.0
astroid==1.6.1
Babel==2.5.3
bcrypt==3.1.4
blinker==1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
Flask==0.12.2
Flask-Admin==1.5.1
Flask-BabelEx==0.9.3
Flask-Dance==0.14.0
Flask-JWT-Extended==3.8.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-RESTful==0.3.6
flask-restplus==0.11.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Social==1.6.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
gevent==1.3.0
graphviz==0.8.3
greenlet==0.4.13
httplib2==0.11.3
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy==1.3
lazy-object-proxy==1.3.1
macholib==1.9
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
msgpack-python==0.5.6
oauth2==1.9.0.post1
oauthlib==2.1.0
passlib==1.7.1
pefile==2017.11.5
pipenv==2018.5.18
prettytable==0.7.2
prompt-toolkit==1.0.15
psycopg2==2.7.4
pycparser==2.18
pydot==1.2.4
Pygments==2.2.0
PyInstaller==3.3.1
PyJWT==1.6.1
pylint==1.8.2
pyparsing==2.2.0
python-crontab==2.3.3
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.4
pyzmq==17.0.0
redis==2.10.6
registry==0.4.2
requests==2.18.4
requests-oauthlib==0.8.0
schedule==0.5.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.6
sqlalchemy-schemadisplay==1.3
SQLAlchemy-Utils==0.33.3
urllib3==1.22
URLObject==2.4.3
virtualenv==15.2.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
zerorpc==0.6.1